Move intent from pkg to model
